Re-did damage calculation to make easier to tweak and balance and clearer as well. fixed pixel gap on main screen added inventory max until selected bug is fixed added all assets started alerts and segues added money labels to most views recalibrated drop chances to make higher tier loot more rare sell price adjustments WIP

+++ Dig Click/OreProtocol.swift
@@ -23,7 +23,7 @@ public enum mountedType: String {
 
 public protocol Rock{
     static var imageSet: [UIImage] {get set}
-    var health: Int { get set }
+    var health: Double { get set }
     var lives: Int { get set }
     var state: Bool { get set}
     var type: rockType {get}
